requirements.txt
-i https://pypi.douban.com/simple/

django
django-simpleui
djangorestframework
drf-yasg
mistune
django-autocomplete-light
django-mdeditor
pillow